requirements.txt
fastapi==0.70.0
pydantic==1.8.2
uvicorn==0.15.0
tidevice==0.5.6

Jinja2>=2.8
Pillow>=3.4.0
requests>=2.11.1
six
mss==4.0.3
numpy<=1.19.3
opencv-contrib-python
facebook-wda>=1.3.3
